ISSUE-26 Repackage to its own instead of ASF package name
