+++ src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,1 +1,1 @@
-org.apache.spark.sql.state.StateStoreDataSourceProvider
+net.heartsavior.spark.sql.state.StateStoreDataSourceProvider
